Reduce code duplication with refactor of params classes. Add check on line object for association
