+++ geoapps/inversion/electricals/direct_current/pseudo_three_dimensions/constants.py
@@ -60,7 +60,6 @@ inversion_defaults = {
     "length_scale_z": 1.0,
     "s_norm": 0.0,
     "x_norm": 2.0,
-    "y_norm": 2.0,
     "z_norm": 2.0,
     "gradient_type": "total",
     "max_irls_iterations": 25,
@@ -320,7 +319,6 @@ app_initializer = {
     "reference_model": 1e-1,
     "s_norm": 0.0,
     "x_norm": 2.0,
-    "y_norm": 2.0,
     "z_norm": 2.0,
     "upper_bound": 100.0,
     "lower_bound": 1e-5,
